requirements.txt
certifi==2020.11.8
chardet==3.0.4
Django==2.1.5
djangorestframework==3.8.2
idna==2.10
pytz==2020.4
requests==2.25.0
urllib3==1.26.2
